Remove value restriction example
While the example is accurate for FIELD_DEFINITIONs and INPUT_FIELD_DEFINITIONs, it actually modifies the schema in dangerous ways that you probably wouldn't want to modify it. For example, once the directive is applied, this scalar is no longer identifiable as a String. It has no relation to a String at all, except that it is parsed and serialized the same way. It is now a `LengthAtMostX`. This makes the schema far less legible and harder to work with since the actual types are now generated dynamically.

It tends to encourage what seems like a dead-end path for input validation. E.g. I tried to expand this to ARGUMENT_DEFINITION so that I could validate inputs. But as soon as I attempt to use a variable to pass into the argument, I got `ValidationError: Variable "$name" of type "String!" used in position expecting type "LengthAtMost42!"`. It's not particularly usable since the type is created dynamically. And of course if you try to change the type to `LengthAtMost42` in the query, you get `ValidationError: Unknown type "LengthAtMost42". Did you mean "String"?`

I think showing this example does more harm than good (at least it did for me) and it'd either be better off removed or altered/expanded into something more production-ready.

I've opened this as a PR, but it's really just a suggestion/feedback with context. I'm relatively new to GraphQL and felt a little betrayed by this example. I still don't know the "right" way to validate length constraints on a String type expect perhaps to manually create an explicit type for each set of restrictions I want.